move alignment repository
