+++ src/alignment/mod.rs
@@ -1,7 +1,11 @@
 mod alignment;
 mod coverage;
 mod read;
+mod repository;
 
 pub use alignment::Alignment;
 pub use coverage::BaseCoverage;
 pub use read::{AlignedRead, RenderingContext, RenderingContextKind, RenderingContextModifier};
+pub use repository::{
+    AlignmentRepository, AlignmentRepositoryEnum, BamRepository, RemoteBamRepository,
+};
